requirements.txt
playwright==1.50.0
playwright-stealth==1.0.6
openpyxl==3.1.5
Pillow==12.1.1
ddgs>=8.1.1
requests==2.32.5
rembg[cpu]>=2.0.50
opencv-python-headless>=4.8.0
numpy>=1.24.0
